Payments and Order Fixes

+++ MyAPI/Core/Entities/CustomerBasket.cs
@@ -21,7 +21,7 @@ namespace Core.Entities
         public int? DeliveryMethodId { get; set; }
         public string ClientSecret { get; set; }
         public string PaymentIntentId { get; set; }
-        public decimal ShippingPrice { get; set; }
+        public double ShippingPrice { get; set; }
 
         public void AddItem(BasketProduct product, int quantity)
         {
